% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Vectors-pkg.R
\docType{methods}
\name{combine,DataFrame,DataFrame-method}
\alias{combine,DataFrame,DataFrame-method}
\title{Combining DataFrame objects}
\usage{
\S4method{combine}{DataFrame,DataFrame}(x, y, ...)
}
\arguments{
\item{x}{A \link[S4Vectors]{DataFrame} object.}

\item{y}{A \link[S4Vectors]{DataFrame} object.}

\item{...}{One or more \link[S4Vectors]{DataFrame} objects.}
}
\value{
A \link[S4Vectors]{DataFrame} object.
}
\description{
Combine multiple \link[S4Vectors]{DataFrame} objects using a union strategy.
}
\details{
Objects are combined based on the \code{row.names} of \code{x},
\code{y}, and \code{...}.
}
\examples{
x <- DataFrame(A = 1:10, B = letters[1:10], row.names = LETTERS[1:10])
y <- DataFrame(B = letters[6:10], C = 10:6, row.names = LETTERS[6:10])
combine(x, y)

}
\author{
Peter Hickey, \email{peter.hickey@gmail.com}
}

