% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicRanges-pkg.R
\docType{methods}
\name{combine,GenomicRanges,GenomicRanges-method}
\alias{combine,GenomicRanges,GenomicRanges-method}
\title{Combining GRanges objects}
\usage{
\S4method{combine}{GenomicRanges,GenomicRanges}(x, y, ...)
}
\arguments{
\item{x}{A \link[GenomicRanges]{GenomicRanges} object.}

\item{y}{A \link[GenomicRanges]{GenomicRanges} object.}

\item{...}{One or more \link[GenomicRanges]{GenomicRanges} objects.}
}
\value{
A \link[GenomicRanges]{GenomicRanges} object.
}
\description{
Combine multiple \link[GenomicRanges]{GenomicRanges} objects.
}
\details{
Combines two or more \link[GenomicRanges]{GenomicRanges} objects
so that the resulting \link[GenomicRanges]{GenomicRanges} object contains
all elements of the original objects. Elements in the returned value are
unique, that is, an element represented in multiple arguments is represented
only only in the result. To perform this operation, \code{combine()} makes
sure that data in shared elements, \strong{including metadata columns
accessible with \code{\link[S4Vectors]{mcols}()}}, are identical in all the
\link[GenomicRanges]{GenomicRanges} objects. Data differences in shared
elements usually cause an error.

\strong{WARNING}: \code{names()} of the \link[GenomicRanges]{GenomicRanges}
inputs are removed prior to combining, hence the returned value has
\code{names()} set to \code{NULL}.
}
\examples{
x <- GRanges(seqnames = c("chr1", "chr2"),
               ranges = IRanges(c(1, 11), c(11, 20)),
               seqinfo = Seqinfo(c("chr1", "chr2", "chr3")))
y <- GRanges(seqnames = c("chr2", "chr3"),
               ranges = IRanges(c(11, 21), c(20, 30)),
               seqinfo = Seqinfo(c("chr1", "chr2", "chr3")))
# x[2] (resp. y[1]) is only included once in the combined object
combine(x, y)

}
\author{
Peter Hickey, \email{peter.hickey@gmail.com}
}

