% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicRanges-pkg.R
\docType{methods}
\name{combine,GRangesList,GRangesList-method}
\alias{combine,GRangesList,GRangesList-method}
\title{Combining GRangesList objects}
\usage{
\S4method{combine}{GRangesList,GRangesList}(x, y, ...)
}
\arguments{
\item{x}{A \link[GenomcRanges]{GRangesList} object.}

\item{y}{A \link[GenomcRanges]{GRangesList} object.}

\item{...}{One or more \link[GenomcRanges]{GRangesList} objects.}
}
\value{
A \link[GenomcRanges]{GRanges} object.
}
\description{
Combine multiple \link[GenomcRanges]{GRangesList} objects using a union
strategy.
}
\details{
Objects are combined based on the \code{names} of \code{x},
\code{y}, and \code{...}.
}
\examples{
x <- GRangesList(
  "a" = GRanges(seqnames = "chr1",
                ranges = IRanges(1, 10),
                seqinfo = Seqinfo(c("chr1", "chr2", "chr3"))),
  "b" = GRanges(seqnames = "chr2",
                ranges = IRanges(11, 20),
                seqinfo = Seqinfo(c("chr1", "chr2", "chr3"))))
# NOTE: y$b is slightly different to x$b
y <- GRangesList(
  "b" = GRanges(seqnames = "chr2",
                ranges = IRanges(12, 21),
                seqinfo = Seqinfo(c("chr1", "chr2", "chr3"))),
  "c" = GRanges(seqnames = "chr3",
                ranges = IRanges(21, 30),
                seqinfo = Seqinfo(c("chr1", "chr2", "chr3"))))
# NOTE: the 'b' element of the combined object include both x$b and y$b
combine(x, y)

}
\author{
Peter Hickey, \url{peter.hickey@gmail.com}
}

