% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Vectors-pkg.R
\docType{methods}
\name{combine,SimpleList,SimpleList-method}
\alias{combine,SimpleList,SimpleList-method}
\title{Combining SimpleList objects}
\usage{
\S4method{combine}{SimpleList,SimpleList}(x, y, ...)
}
\arguments{
\item{x}{A \link[S4Vectors]{SimpleList} object.}

\item{y}{A \link[S4Vectors]{SimpleList} object.}

\item{...}{One or more \link[S4Vectors]{SimpleList} objects.}
}
\value{
A \link[S4Vectors]{SimpleList} object.
}
\description{
Combine multiple \link[S4Vectors]{SimpleList} objects using a union strategy.
}
\details{
Objects are combined \emph{element-wise} (i.e. using
\code{\link[S4Vectors]{mendoapply}}) based on the \code{names} of \code{x},
\code{y}, and \code{...}. This means that \code{x}, \code{y}, and all
elements of \code{...} must be \link[S4Vectors]{SimpleList} objects with the
same length and identical names, and that all elements of each
\link[S4Vectors]{SimpleList} must themselves have well-defined
\code{combine} methods.
}
\examples{
x <- SimpleList(a = DataFrame(A = 1:5, row.names = LETTERS[1:5]),
                b = matrix(1:4, ncol = 2, byrow = TRUE,
                           dimnames = list(letters[1:2], LETTERS[1:2])))
y <- SimpleList(a = DataFrame(A = 3:7, row.names = LETTERS[3:7]),
                b = matrix(1:8, ncol = 2, byrow = TRUE,
                           dimnames = list(letters[1:4], LETTERS[1:2])))
combine(x, y)

}
\author{
Peter Hickey, \email{peter.hickey@gmail.com}
}

