% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment-pkg.R
\docType{methods}
\name{combine,SummarizedExperiment,SummarizedExperiment-method}
\alias{combine,SummarizedExperiment,SummarizedExperiment-method}
\title{Combining SummarizedExperiment/RangedSummarizedExperiment objects}
\usage{
\S4method{combine}{SummarizedExperiment,SummarizedExperiment}(x, y, ...)
}
\arguments{
\item{x}{A \link[S4Vectors]{DataFrame} object.}

\item{y}{A \link[S4Vectors]{DataFrame} object.}

\item{...}{One or more \link[S4Vectors]{DataFrame} objects.}
}
\value{
A \link[S4Vectors]{DataFrame} object.
}
\description{
Combine multiple \link[SummarizedExperiment]{SummarizedExperiment} or
\link[SummarizedExperiment]{RangedSummarizedExperiment} objects using a
union strategy.
}
\details{
\link[SummarizedExperiment]{SummarizedExperiment} objects are
combined based on the \code{names} of \code{x}, \code{y}, and \code{...},
while \link[SummarizedExperiment]{RangedSummarizedExperiment} are combined
based on finding matching \code{rowRanges()} of \code{x}, \code{y}, and
\code{...}. \strong{WARNING}: Does not currently work if the
\code{rowRanges} slot of \code{x}, \code{y}, or \code{...} is a
\link[GenomicRanges]{GRangesList} objects.
}
\section{Note for Developers}{

Any class that extends the
\link[SummarizedExperiment]{SummarizedExperiment} or
\link[SummarizedExperiment]{RangedSummarizedExperiment} class by adding
additional slots will need to be careful when defining a combine() method
for the new class if it wants to call
\code{combine,SummarizedExperiment-method} by inheritance through
\code{callNextMethod()}. Specifically, the \code{combine()} method will need
to first update these additional slots, update the corresponding slots in
\code{x} (thus likely making \code{x} an invalid object), and then calling
\code{callNextMethod()}. An alternative would be to set \code{check = FALSE}
when replacing the slots in the
\code{combine,SummarizedExperiment,SummarizedExperiment-method}, but this
would require that the validity of each slot was checked in some other way
to guard against generally returning of unvalidated
\link[SummarizedExperiment]{SummarizedExperiment} or
\link[SummarizedExperiment]{RangedSummarizedExperiment} objects.
}
\examples{
nrows <- 200; ncols <- 6
counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
colData <- DataFrame(Treatment = rep(c("ChIP", "Input"), 3),
                     row.names = LETTERS[1:6])
se <- SummarizedExperiment(assays = SimpleList(counts = counts),
                            colData = colData)
names(se) <- paste0("f", 1:200)
x <- se[1:150, 1:4]
y <- se[30:170, 2:6]
combine(x, y)
rowRanges <- GRanges(rep(c("chr1", "chr2"), c(50, 150)),
                     IRanges(floor(runif(200, 1e5, 1e6)), width = 100),
                     strand = sample(c("+", "-"), 200, TRUE),
                     feature_id = sprintf("ID\%03d", 1:200))
rse <- se
rownames(rse) <- NULL
rowRanges(rse) <- rowRanges
x <- rse[1:150, 1:4]
y <- rse[30:170, 2:6]
combine(x, y)

}
\author{
Peter Hickey, \email{peter.hickey@gmail.com}
}

