% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment-pkg.R
\docType{methods}
\name{combine,SummarizedExperiment0,SummarizedExperiment0-method}
\alias{combine,SummarizedExperiment0,SummarizedExperiment0-method}
\title{Combining SummarizedExperiment0/RangedSummarizedExperiment objects}
\usage{
\S4method{combine}{SummarizedExperiment0,SummarizedExperiment0}(x, y, ...)
}
\arguments{
\item{x}{A \link[S4Vectors]{DataFrame} object.}

\item{y}{A \link[S4Vectors]{DataFrame} object.}

\item{...}{One or more \link[S4Vectors]{DataFrame} objects.}
}
\value{
A \link[S4Vectors]{DataFrame} object.
}
\description{
Combine multiple \link[SummarizedExperiment]{SummarizedExperiment0} or
\link[SummarizedExperiment]{RangedSummarizedExperiment} objects using a
union strategy.
}
\details{
\link[SummarizedExperiment]{SummarizedExperiment0} objects are
combined based on the \code{names} of \code{x}, \code{y}, and \code{...},
while \link[SummarizedExperiment]{RangedSummarizedExperiment} are combined
based on finding matching genomic ranges. \strong{WARNING}: Does not
currently work if the \code{rowRanges} slot of \code{x}, \code{y}, or
\code{...} is a \link[GenomicRanges]{GRangesList} objets.
}
\examples{
nrows <- 200; ncols <- 6
counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
colData <- DataFrame(Treatment=rep(c("ChIP", "Input"), 3),
                     row.names=LETTERS[1:6])
se0 <- SummarizedExperiment(assays=SimpleList(counts=counts),
                            colData=colData)
names(se0) <- paste0("f", 1:200)
x <- se0[1:150, 1:4]
y <- se0[30:170, 2:6]
combine(x, y)
rowRanges <- GRanges(rep(c("chr1", "chr2"), c(50, 150)),
                     IRanges(floor(runif(200, 1e5, 1e6)), width=100),
                     strand=sample(c("+", "-"), 200, TRUE),
                     feature_id=sprintf("ID\%03d", 1:200))
rse <- unname(se0)
rowRanges(rse) <- rowRanges
x <- se0[1:150, 1:4]
y <- se0[30:170, 2:6]
combine(x, y)

}
\author{
Peter Hickey, \email{peter.hickey@gmail.com}
}

