% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine2.R
\docType{methods}
\name{combine2,SummarizedExperiment0,SummarizedExperiment0-method}
\alias{combine2,SummarizedExperiment0,SummarizedExperiment0-method}
\title{Combining SummarizedExperiment objects}
\usage{
\S4method{combine2}{SummarizedExperiment0,SummarizedExperiment0}(x, y, ...,
  nomatch = NA, use.mcols = FALSE)
}
\description{
Combining SummarizedExperiment objects
}

