% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{densify}
\alias{densify}
\title{Densify sparse assays.}
\usage{
densify(x, i, j, ...)
}
\arguments{
\item{x}{A \link{SparseAssays} object.}

\item{i, j}{Numeric or character vectors indicating which sparse assay
(\code{i}) and samples (\code{j}) to extract and densify. At least
one of \code{i} or \code{j} must be specified; see \sQuote{Details}.}

\item{...}{Additional arguments, for use in specific methods.}
}
\value{
A \code{\link[S4Vectors]{SimpleList}} of length =
        \code{length(i)}, each containing a
        \code{\link[S4Vectors]{SimpleList}} of length = \code{length{j}},
        each containing a \code{matrix} of the densified data for that
        sample in that sparse assay.
}
\description{
A generic function to densify (expand) the elements of a SparseAssays object.
}
\details{
\strong{WARNING}: Since it is generally undesirable to
         simultaneously densify all sparse assays and samples, you must
         specify at least one of \code{i} and \code{j}. If you \emph{really}
         wish to simultaneously densify all sparse assays and samples then
         use \code{densify(x, seq_len(length(x)), seq_len(ncol(x)))}. If
         \code{i} (resp. \code{j}) is missing then effectively
         \code{i = seq_len(length(x))} (resp. \code{j = seq_len(ncol(x))}).
}
\examples{
# See ?SimpleListSparseAssays

}
\author{
Peter Hickey, \url{peter.hickey@gmail.com}
}
\seealso{
\itemize{
 \item \link{SimpleListSparseAssays} for examples.
}
}

