% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangedSparseSummarizedExperiment-class.R
\name{get_rownames_from_sparse_assays}
\alias{get_rownames_from_sparse_assays}
\title{Get rownames of a SparseAssays object}
\usage{
get_rownames_from_sparse_assays(sparse_assays)
}
\arguments{
\item{sparse_assays}{A SimpleListSparseAssays object.}
}
\description{
Get rownames of a SparseAssays object
}

