% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleListSparseAssays-class.R
\docType{class}
\name{SimpleListSparseAssays-class}
\alias{SimpleListSparseAssays}
\alias{SimpleListSparseAssays-class}
\alias{[,SimpleListSparseAssays,ANY-method}
\alias{[<-,SimpleListSparseAssays,ANY,ANY,ANY-method}
\alias{cbind,SimpleListSparseAssays-method}
\alias{combine,SimpleListSparseAssays,SimpleListSparseAssays-method}
\alias{dim,SimpleListSparseAssays-method}
\alias{rbind,SimpleListSparseAssays-method}
\title{SimpleListSparseAssays objects}
\usage{
\S4method{dim}{SimpleListSparseAssays}(x)

\S4method{[}{SimpleListSparseAssays,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{SimpleListSparseAssays,ANY,ANY,ANY}(x, i, j, ...) <- value

\S4method{rbind}{SimpleListSparseAssays}(..., deparse.level = 1)

\S4method{cbind}{SimpleListSparseAssays}(..., deparse.level = 1)

\S4method{combine}{SimpleListSparseAssays,SimpleListSparseAssays}(x, y, ...)
}
\arguments{
\item{x}{A SimpleListSparseAssays object.}

\item{i, j}{Numeric or character vectors indicating which \emph{rows} of the
sparse assays (\code{i}) and samples (\code{j}) to select.}

\item{...}{For \code{cbind()}, \code{rbind()}, and \code{combine()} one or
more SimpleListSparseAssay objects. Otherwise, additional arguments,
for use in specific methods.}

\item{drop}{Not used by \code{[,SimpleListSparseAssays,ANY-method}.}

\item{value}{An object of a class specified in the S4 method signature or as
outlined in \sQuote{Details}.}

\item{deparse.level}{See \code{?base::\link[base]{cbind}} for a description
of this argument.}

\item{y}{A SimpleListSparseAssays object.}
}
\description{
A concrete subclass of the \link{SparseAssays} virtual class.
}
\details{
The SimpleListSparseAssays class has a nested-list structure. This
hierarchy is illustrated below for an example with two sparse assays and
three samples:
\preformatted{
SimpleListSparseAssays
|-- sparse_assay_1
|   |-- sample_1
|   |   |-- key
|   |   |-- value
|   |-- sample_2
|   |   |-- key
|   |   |-- value
|   |-- sample3
|   |   |-- key
|   |   |-- value
|-- sparse_assay_2
|   |-- sample_1
|   |   |-- key
|   |   |-- value
|   |-- sample_2
|   |   |-- key
|   |   |-- value
|   |-- sample_3
|   |   |-- key
|   |   |-- value
}

Each "key" is an integer vector and all key elements must have identical
length. Each "value" element is a matrix object (an assay with a
1-dimensional measurement is stored as a 1-column matrix). Each value
element may have a different number of rows but the maximum number of rows
must be less than or equal to the length of the key elements. A row of the
value element may be pointed to multiple times by the key element within
the same sample and asssay.
}
\section{Dimensions}{

The dimensions of a SimpleListSparseAssays object are defined by nrow =
length of key, and ncol = number of samples.
}

\section{Subsetting}{

Subsetting with \code{[} uses \code{i} to subset \emph{rows} in each
sparse assay and \code{j} to subset samples in each sparse assay. Use
\code{[[} to select the \code{i}-th sparse assay.
}

\section{Combining}{

SimpleListSparseAssays objects can be combined in three different ways.
\enumerate{
 \item \code{rbind} Suitable for when each object has the same samples.
 \item \code{cbind} Suitable for when each object has unique samples.
 \item \code{combine} Suitable in either case, \strong{however}, requires
 that \code{dimnames} are set on each object and that all objects have an
 identical number of sparse assays with identical names.
}
}

\section{Densifying and coercion}{

SimpleListSparseAssays objects can densified using the
\code{\link{densify}} method. It is generally advisable to only
densify a single sample and an sparse assay at a time; see
\code{\link{densify}} for details.

SimpleListSparseAssays objects can also coerced
into a
\link[SummarizedExperiment]{ShallowSimpleListAssays} object (from the
\pkg{SummarizedExperiment} package); this will also densify the object.
This can be done using \code{as(x, "ShallowSimpleListAssays")}, where
\code{x} is a SimpleListSparseAssays object. \emph{WARNING}: The resulting
\link[SummarizedExperiment]{ShallowSimpleListAssays} object will typically
require much more memory than the equivalent SimpleListSparseAssays object.
}
\examples{
# TODO: Get old examples from docs
## ---------------------------------------------------------------------
## DIRECT MANIPULATION OF SparseAssays OBJECTS
## ---------------------------------------------------------------------
sl1 <- SimpleList(
  s1 = SimpleList(key = as.integer(c(NA, 1, NA, NA, 2, NA, 3, NA, 4, 5)),
                  value = matrix(1:10, ncol = 2)),
  s2 = SimpleList(key = as.integer(c(NA, NA, 1, 2, NA, NA, 3, 4, NA, NA)),
                  value = matrix(8:1, ncol = 2)))

sl2 <- SimpleList(
  s1 = SimpleList(key = as.integer(c(NA, 1, NA, 2, 2, NA, 1, NA, NA, 1)),
                  value = matrix(1:2, ncol = 1)),
  s2 = SimpleList(key = as.integer(c(1, 1, 1, 2, NA, NA, NA, NA, NA, NA)),
                  value = matrix(4:3, ncol = 1)))
sa <- SparseAssays(SimpleList(sl1, sl2))
sa

as(sa, "SimpleList")

length(sa)
sa[[2]]
dim(sa)

sa2 <- sa[-4, 2]
sa2
length(sa2)
sa2[[2]]
dim(sa2)

names(sa)
names(sa) <- c("sa1", "sa2")
names(sa)
sa[["sa2"]]

rbind(sa, sa)
\dontrun{
  # ERROR: cbind-ing requires unique sample names
  cbind(sa, sa)
}
\dontrun{
  # ERROR: missing dimnames (which can't because there is no
  #        dimnames,SparseAssays-method.
  combine(sa[1:7, 1], sa[3:10, 2])
}

}
\author{
Peter Hickey, \email{peter.hickey@gmail.com}
}
\seealso{
\itemize{
 \item \link{SparseAssays} objects for additional methods.
}
}

