% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{sparsify}
\alias{sparsify}
\alias{sparsify,data.frame,character-method}
\alias{sparsify,data.table,character-method}
\alias{sparsify,matrix,character-method}
\title{Sparsify a matrix-like object.}
\usage{
sparsify(x, return_class, ...)
}
\arguments{
\item{x}{A matrix-like object, e.g., \code{\link[base]{matrix}},
\code{\link[base]{data.frame}}, or
\code{\link[data.table]{data.table}}.}

\item{return_class}{The class of the returned object.}

\item{...}{Additional arguments, for use in specific methods.}
}
\value{
An object with the concepts of a \code{key} and a \code{value}. The
specific return value is left to individual methods. For example, the
\code{sparsify,SimpleList,matrix-method} returns a
\link[S4Vectors]{SimpleList} object with an \code{integer} \code{key}
element and a \code{matrix} \code{value} element.
}
\description{
\strong{WARNING}: This is intended for developers and not end-users; see
\sQuote{Note for Developers}.
}
\details{
A generic function to sparsify a matrix-like object, e.g.,
\code{\link[base]{matrix}}, \code{\link[base]{data.frame}}, or
\code{\link[data.table]{data.table}}.

While the concepts of "sparsify" and "densify" are inverse
operations, in general, \code{identical(densify(sparsify(x)), x)} is
\strong{\code{FALSE}}. This is because the return value of \code{sparsify()}
is generally not compatible with the signature of a
\code{\link{densify}()} method.
}
\section{Note for Developers}{

This generic and method are used internally by the
\pkg{SparseSummarizedExperiment} package but are not intended for use by
end-users. End-users with their data as dense \code{\link[base]{matrix}}
objects should use the \code{\link{SparseAssays}} constructor in conjunction
with the associated \code{\link{combine}} method to sparify their data and
return a valid SparseAssays object.

One particular word of warning, the 'value' element returned by
\code{sparsify} may contain the \code{NA}-row and will not have \code{NA}s
in the 'key' element.
}
\examples{
# See ?SimpleListSparseAssays

}
\author{
Peter Hickey, \email{peter.hickey@gmail.com}
}
\seealso{
\itemize{
 \item \link{SimpleListSparseAssays}
}
}
\keyword{internal}

