% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicRanges-pkg.R
\docType{methods}
\name{combine,GenomicRanges,GenomicRanges-method}
\alias{combine,GenomicRanges,GenomicRanges-method}
\title{Combining GRanges objects}
\usage{
\S4method{combine}{GenomicRanges,GenomicRanges}(x, y, ...)
}
\arguments{
\item{x}{A \link[GenomicRanges]{GenomicRanges} object.}

\item{y}{A \link[GenomicRanges]{GenomicRanges} object.}

\item{...}{One or more \link[GenomicRanges]{GenomicRanges} objects.}
}
\value{
A \link[GenomicRanges]{GenomicRanges} object.
}
\description{
Combine multiple \link[GenomicRanges]{GenomicRanges} objects.
}
\details{
If and only if all of \code{x}, \code{y}, and \code{...} have
non-\code{NULL} \code{names()}, then objects are combined using these names.
Otherwise, names are first created based on finding identical ranges (using
\code{\link[S4Vectors]{findMatches}()}) and then the objects are combined
using these names. All of \code{x}, \code{y}, and \code{...} must also have
compatible \code{\link[S4Vectors]{mcols}()}.
}
\examples{
x <- GRanges(seqnames = c("chr1", "chr2"),
               ranges = IRanges(c(1, 11), c(11, 20)),
               seqinfo = Seqinfo(c("chr1", "chr2", "chr3")))
y <- GRanges(seqnames = c("chr2", "chr3"),
               ranges = IRanges(c(11, 21), c(20, 30)),
               seqinfo = Seqinfo(c("chr1", "chr2", "chr3")))
combine(x, y)

}
\author{
Peter Hickey, \email{peter.hickey@gmail.com}
}

