% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangedSparseSummarizedExperiment-class.R
\docType{class}
\name{RangedSparseSummarizedExperiment-class}
\alias{RangedSparseSummarizedExperiment}
\alias{RangedSparseSummarizedExperiment-class}
\alias{SparseSummarizedExperiment,SparseAssays-method}
\alias{SparseSummarizedExperiment,missing-method}
\title{RangedSparseSummarizedExperiment objects}
\usage{
\S4method{SparseSummarizedExperiment}{SparseAssays}(sparseAssays,
  assays = SimpleList(), rowRanges = GRangesList(), colData = DataFrame(),
  metadata = list())

\S4method{SparseSummarizedExperiment}{missing}(sparseAssays, ...)
}
\arguments{
\item{sparseAssays}{A \link{SparseAssays} object. All dimension names (if
present) must be consistent across elements and with the row names of
\code{rowRanges} and \code{colData}.}

\item{assays}{A \code{list} or \code{SimpleList} of matrix elements,
    or a \code{matrix}. All elements of the list must have the same
    dimensions, and dimension names (if present) must be consistent
    across elements and with the row names of \code{rowRanges} and
    \code{colData}.}

\item{rowRanges}{A \link{GRanges} or \link{GRangesList} object describing 
    the ranges of interest. Names, if present, become the row names of the
    RangedSummarizedExperiment. The length of the \link{GRanges} or the
    \link{GRangesList} must equal the number of rows of the matrices in
    \code{assays}. If \code{rowRanges} is missing, a
    \link{SummarizedExperiment0} instance is returned.}

\item{colData}{An optional \link{DataFrame} describing the
    samples. Row names, if present, become the column names of the
    RangedSummarizedExperiment.}

\item{metadata}{An optional \code{list} of arbitrary content
    describing the overall experiment.}

\item{...}{For \code{SummarizedExperiment}, S4 methods \code{list}
    and \code{matrix}, arguments identical to those of the
    \code{SimpleList} method.

    For \code{rowRanges}, ignored.

  }
}
\description{
RangedSparseSummarizedExperiment objects
}

