% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RangedSparseSummarizedExperiment-class.R
\docType{class}
\name{RangedSparseSummarizedExperiment-class}
\alias{RangedSparseSummarizedExperiment-class}
\alias{SparseSummarizedExperiment,SparseAssays-method}
\alias{SparseSummarizedExperiment,missing-method}
\alias{[,RangedSparseSummarizedExperiment,ANY-method}
\alias{[<-,RangedSparseSummarizedExperiment,ANY,ANY,RangedSparseSummarizedExperiment-method}
\alias{cbind,RangedSparseSummarizedExperiment-method}
\alias{combine2,RangedSparseSummarizedExperiment,RangedSparseSummarizedExperiment-method}
\alias{rbind,RangedSparseSummarizedExperiment-method}
\alias{show,RangedSparseSummarizedExperiment-method}
\alias{sparseAssay,RangedSparseSummarizedExperiment,character-method}
\alias{sparseAssay,RangedSparseSummarizedExperiment,missing-method}
\alias{sparseAssay,RangedSparseSummarizedExperiment,numeric-method}
\alias{sparseAssayNames,RangedSparseSummarizedExperiment-method}
\alias{sparseAssayNames<-,RangedSparseSummarizedExperiment-method}
\alias{sparseAssays,RangedSparseSummarizedExperiment-method}
\alias{sparseAssays<-,RangedSparseSummarizedExperiment-method}
\title{RangedSparseSummarizedExperiment objects}
\usage{
\S4method{SparseSummarizedExperiment}{SparseAssays}(sparseAssays,
  assays = SimpleList(), rowRanges = GRangesList(), colData = DataFrame(),
  metadata = list())

\S4method{SparseSummarizedExperiment}{missing}(sparseAssays, ...)

\S4method{sparseAssays}{RangedSparseSummarizedExperiment}(x, ...,
  withDimnames = TRUE, expand = FALSE)

\S4method{sparseAssays}{RangedSparseSummarizedExperiment}(x, ...,
  withDimnames = TRUE) <- value

\S4method{sparseAssay}{RangedSparseSummarizedExperiment,missing}(x, i, ...,
  withDimnames = TRUE, expand = FALSE)

\S4method{sparseAssay}{RangedSparseSummarizedExperiment,numeric}(x, i, ...,
  withDimnames = TRUE, expand = FALSE)

\S4method{sparseAssay}{RangedSparseSummarizedExperiment,character}(x, i, ...,
  withDimnames = TRUE, expand = FALSE)

\S4method{sparseAssayNames}{RangedSparseSummarizedExperiment}(x, ...)

\S4method{sparseAssayNames}{RangedSparseSummarizedExperiment}(x, ...) <- value

\S4method{[}{RangedSparseSummarizedExperiment,ANY}(x, i, j, ..., drop = TRUE)


  \S4method{[}{RangedSparseSummarizedExperiment,ANY,ANY,RangedSparseSummarizedExperiment}(x,
  i, j, ...) <- value

\S4method{show}{RangedSparseSummarizedExperiment}(object)

\S4method{rbind}{RangedSparseSummarizedExperiment}(..., deparse.level = 1)

\S4method{cbind}{RangedSparseSummarizedExperiment}(..., deparse.level = 1)


  \S4method{combine2}{RangedSparseSummarizedExperiment,RangedSparseSummarizedExperiment}(x,
  y, ..., nomatch = NA, use.mcols = FALSE)
}
\description{
RangedSparseSummarizedExperiment objects
}
\examples{
sa <- SparseAssays(sparse_assays =
                     SimpleList(a1 =
                                  SimpleList(
                                    s1 = SimpleList(map =
                                                      as.integer(c(NA, 1, NA, NA, 2, NA, 3, NA, 4, 5)),
                                                    data =
                                                      matrix(1:10, ncol = 2)),
                                    s2 = SimpleList(map =
                                                      as.integer(c(NA, NA, 1, 2, NA, NA, 3, 4, NA, NA)),
                                                    data =
                                                      matrix(8:1, ncol = 2))),
                                a2 =
                                  SimpleList(
                                    s1 = SimpleList(map =
                                                      as.integer(c(NA, 1, NA, 2, 2, NA, 1, NA, NA, 1)),
                                                    data = matrix(1:2, ncol = 1)),
                                    s2 = SimpleList(map =
                                                      as.integer(c(1, 1, 1, 2, NA, NA, NA, NA, NA, NA)),
                                                    data = matrix(4:3, ncol = 1)))

                     )
)
rse <- SummarizedExperiment(rowRanges = GRanges('chr1', IRanges(1:10, 2:11)),
                            colData = DataFrame(row.names = c("s1", "s2")))
rsse <- SparseSummarizedExperiment(sparseAssays = sa,
                                   rowRanges = rowRanges(rse),
                                   colData = colData(rse))
x <- rsse
y <- new("RangedSparseSummarizedExperiment")
}

