% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleListSparseAssays-class.R
\docType{class}
\name{SimpleListSparseAssays-class}
\alias{SimpleListSparseAssays}
\alias{SimpleListSparseAssays-class}
\alias{[,SimpleListSparseAssays,ANY-method}
\alias{[<-,SimpleListSparseAssays,ANY,ANY,ANY-method}
\alias{cbind,SimpleListSparseAssays-method}
\alias{combine,SimpleListSparseAssays,SimpleListSparseAssays-method}
\alias{dim,SimpleListSparseAssays-method}
\alias{rbind,SimpleListSparseAssays-method}
\title{SimpleListSparseAssays objects}
\usage{
\S4method{dim}{SimpleListSparseAssays}(x)

\S4method{[}{SimpleListSparseAssays,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{SimpleListSparseAssays,ANY,ANY,ANY}(x, i, j, ...) <- value

\S4method{rbind}{SimpleListSparseAssays}(..., deparse.level = 1)

\S4method{cbind}{SimpleListSparseAssays}(..., deparse.level = 1)

\S4method{combine}{SimpleListSparseAssays,SimpleListSparseAssays}(x, y, ...)
}
\arguments{
\item{x}{A SimpleListSparseAssays object.}

\item{i, j}{Numeric or character vectors indicating which \emph{rows} of the
sparse assays (\code{i}) and samples (\code{j}) to select.}

\item{...}{For \code{cbind()}, \code{rbind()}, and \code{combine()} one or
more SimpleListSparseAssay objects. Otherwise, additional arguments,
for use in specific methods.}

\item{drop}{Not used by \code{[,SimpleListSparseAssays,ANY-method}.}

\item{value}{An object of a class specified in the S4 method signature or as
outlined in \sQuote{Details}.}

\item{deparse.level}{See \code{?base::\link[base]{cbind}} for a description
of this argument.}

\item{y}{A SimpleListSparseAssays object.}
}
\description{
A concrete subclass of the \link{SparseAssays} virtual class.
As such, all the methods documented in
\code{?}\link{SparseAssays} also work on a SimpleListSparseAssays object.
See \link{SparseAssays} for details.
}
\details{
The SimpleListSparseAssays class has a nested-list structure. This
hierarchy is illustrated below for an example with two sparse assays and
three samples:
\preformatted{
SimpleListSparseAssays
|-- sparse_assay_1
|   |-- sample_1
|   |   |-- key
|   |   |-- value
|   |-- sample_2
|   |   |-- key
|   |   |-- value
|   |-- sample_3
|   |   |-- key
|   |   |-- value
|-- sparse_assay_2
|   |-- sample_1
|   |   |-- key
|   |   |-- value
|   |-- sample_2
|   |   |-- key
|   |   |-- value
|   |-- sample_3
|   |   |-- key
|   |   |-- value
}

Each \sQuote{key} is an integer vector and all key elements must have
identical length. Each \sQuote{value} element is a matrix object. Each
value element may have a different number of rows but the maximum number of
rows must be less than or equal to the length of the key elements. A row of
the value element may be pointed to multiple times by the key element
within the same sample and asssay.
}
\examples{
# TODO: Get old examples from docs?
# TODO: Use SparseAssays(matrix) in examples?
# TODO: Set names on 'key' elements so that can demonstrate combine().
## ---------------------------------------------------------------------
## DIRECT MANIPULATION OF SparseAssays OBJECTS
## ---------------------------------------------------------------------
sl1 <- SimpleList(
  s1 = SimpleList(key = as.integer(c(NA, 1, NA, NA, 2, NA, 3, NA, 4, 5)),
                  value = matrix(1:10, ncol = 2)),
  s2 = SimpleList(key = as.integer(c(NA, NA, 1, 2, NA, NA, 3, 4, NA, NA)),
                  value = matrix(8:1, ncol = 2)))

sl2 <- SimpleList(
  s1 = SimpleList(key = as.integer(c(NA, 1, NA, 2, 2, NA, 1, NA, NA, 1)),
                  value = matrix(1:2, ncol = 1)),
  s2 = SimpleList(key = as.integer(c(1, 1, 1, 2, NA, NA, NA, NA, NA, NA)),
                  value = matrix(4:3, ncol = 1)))
sa <- SparseAssays(SimpleList(sa1 = sl1, sa2 = sl2))
sa

as(sa, "SimpleList")

length(sa)
sa[[2]]
dim(sa)

sa2 <- sa[-4, 2]
sa2
length(sa2)
sa2[[2]]
dim(sa2)

names(sa)
names(sa) <- c("sa1", "sa2")
names(sa)
sa[["sa2"]]

rbind(sa, sa)
\dontrun{
  # ERROR: cbind-ing requires unique sample names
  cbind(sa, sa)
}
# Works because unique sample names
cbind(sa[, 1], sa[, 2])
\dontrun{
  # ERROR: combine() requires named 'key' elements
  combine(sa[1:7, 1], sa[3:8, 1])
  # ERROR: combine() requires named 'key' elements
  combine(sa[1:7, 1], sa[3:8, 2])
}
# TODO: An example of combine() that works because keys are named.

}
\author{
Peter Hickey, \email{peter.hickey@gmail.com}
}
\seealso{
\itemize{
 \item \link{SparseAssays} objects for a description of the available
   methods.
}
}

