% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseAssays-class.R
\docType{class}
\name{SparseAssays-class}
\alias{NROW,SparseAssays-method}
\alias{SparseAssays}
\alias{SparseAssays-class}
\alias{[[,SparseAssays,ANY,ANY-method}
\alias{[[<-,SparseAssays,ANY,ANY-method}
\alias{length,SparseAssays-method}
\alias{names,SparseAssays-method}
\alias{names<-,SparseAssays-method}
\title{SparseAssays objects}
\usage{
SparseAssays(sparse_assays = SimpleList(), subclass)

\S4method{length}{SparseAssays}(x)

\S4method{NROW}{SparseAssays}(x)

\S4method{names}{SparseAssays}(x)

\S4method{names}{SparseAssays}(x) <- value

\S4method{[[}{SparseAssays,ANY,ANY}(x, i, j, ...)

\S4method{[[}{SparseAssays,ANY,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{sparse_assays}{A SimpleList or list of sparse assays.}

\item{subclass}{The concrete subclass to be instantiated. The default is
\link{SimpleListSparseAssays}.}

\item{x}{A SparseAssays object.}

\item{value}{An object of a class specified in the S4 method signature or as
outlined in \sQuote{Details}.}

\item{i}{Numeric or character vector of length 1 indicating which sparse
assay to select.}

\item{j}{Not used by \code{[[,SparseAssays,ANY,ANY-method} or
\code{[[<-,SparseAssays,ANY,ANY-method}.}

\item{...}{Additional arguments, for use in specific methods.}
}
\value{
For \code{SparseAssays()}, an instance of a concrete SparseAssays
subclass given by \code{subclass}.
}
\description{
The SparseAssays virtual class and its methods provide a
formal abstraction of the sparseAssays slot of
\link{SparseSummarizedExperiment} and
\link{RangedSparseSummarizedExperiment} objects.

\link{SimpleListSparseAssays} and SimpleListJointSparseAssays (\strong{not
yet implemented}) are concrete subclasses of SparseAssays with the former
being currently the default implementation of SparseAssays objects. Other
implementations (e.g. disk-based, environment-based) could easily be added.

Note that these classes are not meant to be used directly by the end-user
and the material in this man page is aimed at package developers.
}
\details{
SparseAssays objects have a list-like semantics with elements
containing key and value elements.

The SparseAssays API consists of:
\itemize{
 \item (a) The \code{SparseAssays()} constructor function.
 \item (b) Lossless back and forth coercion from/to
 \code{\link[S4Vectors]{SimpleList}}. The coercion
 method from \code{\link[S4Vectors]{SimpleList}} doesn't need (and should
 not) validate the returned object.
 \item (c) \code{\link{length}}, \code{\link{NROW}}, \code{\link{names}},
 \code{\link{names<-}}, \code{\link{[[}}, \code{\link{[[<-}}
 \item (d) \code{\link{dim}}, \code{\link{[}}, \code{\link{[<-}},
 \code{\link{rbind}}, \code{\link{cbind}}, \code{\link{combine}},
 \code{\link{densify}}
}

A SparseAssays concrete subclass needs to implement (b) (required) plus
the methods in (d) (required). The methods in (c) are inherited from the
\code{\link[S4Vectors]{SimpleList}} class. Each element of a SparseAssays
object is referred to as a "sparse assay" (lowercase).

\strong{IMPORTANT}: Methods that return a modified SparseAssays object
(a.k.a. endomorphisms), that is, \code{[} as well as replacement methods
\code{names<-}, \code{[[<-}, and \code{[<-}, must respect the
\emph{copy-on-change contract}.
With objects that don't make use of references internally, the developer
doesn't need to take any special action for that because it's automatically
taken care of by R itself. However, for objects that do make use of
references internally (e.g. environments, external pointers, pointer to a
file on disk, etc...), the developer needs to be careful to implement
endomorphisms with copy-on-change semantics. This can easily be achieved by
performaing a full (deep) copy of the object before modifying it instead of
trying to modify it in-place. Note that the full (deep) copy is not always
necessary in order to achieve copy-on-change semantics: it's enough (and
often preferrable for performance reasons) to copy only the parts of the
objects that need to be modified.

SparseAssays has currently 1 implementation formalized by concrete subclass
\link{SimpleListSparseAssays}. There are written specs for a second
formalization,  SimpleListJointSparseAssays, although this is not yet
implemented.

The sparseAssays slot of a \link{SparseSummarizedExperiment} object contains
an instance of \link{SimpleListSparseAssays}.

\strong{NOTE}: SparseAssays only payoff compared to
\code{SummarizedExperiment::\link[SummarizedExperiment]{Assays}} when you
get more than one measurement per-feature, per-sample. The payoff is greater
when there are lots of features with the same measurement (normally within a
sample, although SimpleListJointSparseAssays should allow this constraint to
be removed) and/or lots of NAs per-sample.
}
\examples{
# See ?SimpleListSparseAssays

}
\author{
Peter Hickey, \email{peter.hickey@gmail.com}
}
\seealso{
\itemize{
 \item \link{SparseSummarizedExperiment} objects.
 \item \link[S4Vectors]{SimpleList} objects in the \pkg{S4Vectors} package.
}
}

