% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SparseSummarizedExperiment-class.R
\docType{class}
\name{SparseSummarizedExperiment-class}
\alias{SparseSummarizedExperiment-class}
\alias{[,SparseSummarizedExperiment,ANY-method}
\alias{[<-,SparseSummarizedExperiment,ANY,ANY,SparseSummarizedExperiment-method}
\alias{cbind,SparseSummarizedExperiment-method}
\alias{combine,SparseSummarizedExperiment,SparseSummarizedExperiment-method}
\alias{rbind,SparseSummarizedExperiment-method}
\alias{show,SparseSummarizedExperiment-method}
\alias{sparseAssay,SparseSummarizedExperiment,character-method}
\alias{sparseAssay,SparseSummarizedExperiment,missing-method}
\alias{sparseAssay,SparseSummarizedExperiment,numeric-method}
\alias{sparseAssayNames,SparseSummarizedExperiment-method}
\alias{sparseAssayNames<-,SparseSummarizedExperiment-method}
\alias{sparseAssays,SparseSummarizedExperiment-method}
\alias{sparseAssays<-,SparseSummarizedExperiment-method}
\title{SparseSummarizedExperiment objects}
\usage{
\S4method{sparseAssays}{SparseSummarizedExperiment}(x, ...,
  withDimnames = TRUE, expand = FALSE)

\S4method{sparseAssays}{SparseSummarizedExperiment}(x, ...,
  withDimnames = TRUE) <- value

\S4method{sparseAssay}{SparseSummarizedExperiment,missing}(x, i, ...,
  withDimnames = TRUE, expand = FALSE)

\S4method{sparseAssay}{SparseSummarizedExperiment,numeric}(x, i, ...,
  withDimnames = TRUE, expand = FALSE)

\S4method{sparseAssay}{SparseSummarizedExperiment,character}(x, i, ...,
  withDimnames = TRUE, expand = FALSE)

\S4method{sparseAssayNames}{SparseSummarizedExperiment}(x, ...)

\S4method{sparseAssayNames}{SparseSummarizedExperiment}(x, ...) <- value

\S4method{[}{SparseSummarizedExperiment,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{SparseSummarizedExperiment,ANY,ANY,SparseSummarizedExperiment}(x,
  i, j, ...) <- value

\S4method{show}{SparseSummarizedExperiment}(object)

\S4method{rbind}{SparseSummarizedExperiment}(..., deparse.level = 1)

\S4method{cbind}{SparseSummarizedExperiment}(..., deparse.level = 1)

\S4method{combine}{SparseSummarizedExperiment,SparseSummarizedExperiment}(x, y,
  ...)
}
\description{
SparseSummarizedExperiment objects
}

